requirements.txt
# Trading Bot Dependencies
# Install: pip install -r requirements.txt

# ======================
# Web Framework
# ======================
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.6

# ======================
# Data & Analysis
# ======================
pandas==2.2.0
numpy==1.26.3
ta==0.11.0

# ======================
# Visualization (for backtest)
# ======================
matplotlib==3.8.2

# ======================
# MT5 Connection (Windows only)
# ======================
MetaTrader5==5.0.45

# ======================
# Configuration
# ======================
python-dotenv==1.0.0

# ======================
# Optional: Telegram Notifications
# ======================
# python-telegram-bot==20.7
